requirements.txt
certifi==2019.6.16
chardet==3.0.4
Django==2.2.2
django-filter==2.1.0
djangorestframework==3.9.4
httpie==1.0.2
idna==2.8
Markdown==3.1.1
Pygments==2.4.2
pytz==2019.1
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.3
